WK 12 Task 1 Video
Video for Task 1 of week 12 submission.

There is a small disclaimer in the beginning, and that in my UNity editor to test my gameplay, it freezes a lot because I have limited disk space. However, I am still able to record the code in action, just with some pauses in between.